move unique name to base class
