+++ asapdiscovery-data/asapdiscovery/data/schema_v2/sets.py
@@ -53,6 +53,14 @@ class MultiStructureBase(BaseModel):
     def __neq__(self, other: Any) -> bool:
         return not self.__eq__(other)
 
+    def unique_name(self):
+        import hashlib
+        data = ""
+        for c in self.complexes:
+            data += c.hash()
+        complexes_hash = hashlib.sha256(data.encode()).hexdigest()
+        return f"{self.ligand.compound_name}-{self.ligand.fixed_inchikey}_{complexes_hash}"
+
 
 class CompoundMultiStructure(MultiStructureBase):
     """
